Added Constants
Added SPEED_MULTIPLIER and ACCELERATION_MULTIPIER constants. Speed
multiplier allows us to set the top speed lower (this would allow for
greater control, which could be useful considering the arena is small).
Acceleration multiplier doesn't do anything yet. The idea was to change
the speed by a fraction of the goal speed until it met the goal speed
providing smoother acceleration. The advantage of this is we may be able
to make it smoother without having to mess with elapsed time.